Apply ESP32 changes
Apply ESP32 related changes from
https://github.com/mattkj/super-go-play on top of gnuboy sources.

+++ pcm.h
@@ -4,12 +4,13 @@
 
 
 #include "defs.h"
+#include <stdint.h>
 
 struct pcm
 {
 	int hz, len;
 	int stereo;
-	byte *buf;
+	int16_t* buf;
 	int pos;
 };
 
@@ -17,5 +18,3 @@ extern struct pcm pcm;
 
 
 #endif
-
-
